software/smapper_toolbox/pyproject.toml
[project]
name = "smapper-toolbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coloredlogs>=15.0.1",
    "docker>=7.1.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0.2",
    "rosbags>=0.10.10",
    "typer>=0.15.4",
]
